paper_reproductions/[2-08]_Vessel_Segmentation/requirements.txt
# 小波框架血管分割依赖
# Python >= 3.8

# 深度学习框架
torch>=1.10.0
torchvision>=0.11.0

# 核心数值计算
numpy>=1.20.0
scipy>=1.7.0

# 小波变换
PyWavelets>=1.1.0

# 图像处理
scikit-image>=0.18.0
opencv-python>=4.5.0
Pillow>=8.0.0

# 可视化
matplotlib>=3.3.0
seaborn>=0.11.0

# 数据处理
tqdm>=4.60.0
pandas>=1.3.0

# 进度条和日志
loguru>=0.5.0

# 可选: TensorBoard
# tensorboard>=2.8.0

# 可选: Weights & Biases
# wandb>=0.12.0
